feat: add Nix flake
Maintaining Nix flakes in-tree will make it easier for NixOS or Nix users to
contribute to Bevy. The alternatives are to either install all Bevy dependencies
globally (which defeats the point of Nix) or to have each user maintain their
own Nix flake. The latter is somewhat onerous because flakes must be tracked
by git; so users must resort to tedious workarounds to use their flake without
actually committing it to the Bevy repo.

The added flake will install a version of the Rust toolchain as configured by
`rust-toolchain.toml`. It also install several libraries that Bevy needs to run
on Linux.

Another benefit of maintaining a flake in-tree is that all Nix users can
contribute to make the flake better, and it should be possible to make sure the
flake continues to work in CI as Bevy adopts more dependencies.